pyproject.toml
[project]
name = "grad-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Django>=5.0.2",
    "djangorestframework>=3.15.2",
    "django-cors-headers>=4.3.1",
    "django-filter>=24.1",
    "django-extensions>=3.2.3",
    "django-environ>=0.11.2",
    "django-debug-toolbar>=4.3.0",
    "django-rest-swagger>=2.2.0",
    "drf-yasg>=1.21.7",
    "psycopg2-binary>=2.9.9",
    "celery>=5.3.4",
    "redis>=5.0.1",
    "channels>=4.0.0",
    "channels-redis>=4.0.0",
    "django-chunked-upload>=2.0.0",
    "requests>=2.32.3",
    "httpx>=0.27.0",
    "langchain-core>=0.3.62",
    "langsmith>=0.3.43",
    "langgraph-checkpoint>=2.0.26",
    "langgraph-checkpoint-sqlite>=2.0.10",
    "sqlite-vec>=0.1.6",
    "pydantic>=2.11.5",
    "pydantic_core>=2.33.2",
    "typing_extensions>=4.13.2",
    "typing-inspection>=0.4.1",
    "Jinja2>=3.1.4",
    "attrs>=23.2.0",
    "pillow>=10.3.0",
    "python-magic>=0.4.27",
    "langid>=1.1.6",
    "langgraph>=0.6.3",
    "langchain-google-genai>=2.1.9",
    "dotenv>=0.9.9",
    "langchain>=0.3.27",
    "pandas>=2.3.1",
    "langdetect>=1.0.9",
    "ipython>=9.4.0",
    "rapidfuzz>=3.13.0",
    "faker>=37.5.3",
    "djangorestframework-simplejwt>=5.5.1",
    "daphne>=4.2.1",
    "ebooklib>=0.19",
]
